Update patch set 1
Patch Set 1: Fails; Do not submit

The names and locations of the files are intentional. These files are meant to be pulled from a CM9 ROM, not a stock ROM, as a large portion of them came from other devices, as well as other reasons. A ROM compiled with files pulled after using these changes will not work.

Patch-set: 1